requirements.txt
opencv-python
mediapipe
numpy
python-dotenv
google-generativeai
